Move is_same check from checks to utils
Semantics and UX-wise it's not really a check
