+++ pontoon/checks/libraries/pontoon.py
@@ -19,29 +19,6 @@ def get_max_length(comment):
     return None
 
 
-def is_same(same_translations, can_translate):
-    """
-    Check if translation is the same
-    :arg QuerySet `same_translations`: translations that have the same string
-        as a suggestion/translation.
-    :arg boolean `can_translate`: user is able to submit translations
-    :returns: True if translation already exists and won't be submitted to the database.
-    """
-    if not same_translations:
-        return False
-
-    st = same_translations[0]
-
-    if can_translate:
-        if st.approved and not st.fuzzy:
-            return True
-
-    else:
-        if not st.fuzzy:
-            return True
-    return False
-
-
 def run_checks(entity, string):
     """
     Group all checks related to the base UI
@@ -69,4 +46,5 @@ def run_checks(entity, string):
         checks['pErrors'].append(
             'Newline characters are not allowed.'
         )
+
     return checks
